Refactor file structure and create security rules enum
